refactor: split VideoGen.jsx into hooks + subcomponents (#2834)
